requirements.ide.txt
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
attrs==23.1.0
autoflake==2.2.0
Automat==22.10.0
bcrypt==4.0.1
beanie==1.20.0
certifi==2023.7.22
click==8.1.6
constantly==15.1.0
coverage==7.2.7
decorator==5.1.1
dill==0.3.7
dnspython==2.4.1
docstring-to-markdown==0.12
doq==0.9.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-model==0.0.5
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
motor==3.2.0
orjson==3.9.2
outcome==1.2.0
packaging==23.1
parso==0.8.3
passlib==1.7.4
platformdirs==3.10.0
pluggy==1.2.0
pyasn1==0.5.0
pydantic==1.10.12
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyflakes==3.0.1
pylint==2.17.5
pymongo==4.4.1
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.0
python-dotenv==1.0.0
python-jose==3.3.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
trio==0.22.2
Twisted==22.10.0
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
yapf==0.40.1
zipp==3.16.2
zope.interface==6.0
